+++ src/common/components/ListFile/ListFileSnsEventAssistant/ListFileSnsEventAssistant.client.tsx
@@ -2,8 +2,8 @@
 
 import { FeaturedFileIconsState } from '@icons/FeaturedFileIcons/FeaturedFileIcons.types';
 
-import BaseListFile from '../BaseListFile/BaseListFile';
-import HumanIconWithText from '../HumanIconWithText/HumanIconWithText';
+import BaseListFile from '../BaseListFile/BaseListFile.client';
+import HumanIconWithText from '../HumanIconWithText/HumanIconWithText.server';
 import { ListFileSnsEventAssistantProps } from './ListFileSnsEventAssistant.types';
 
 const ListFileSnsEventAssistant = ({
